pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pywhois2"
version = "0.1.0"
description = "WHOIS lookup and parsing library with CLI support"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.8"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "blacknon", email = "blacknon@orebibou.com" },
]
maintainers = [
  { name = "blacknon", email = "blacknon@orebibou.com" },
]
dependencies = [
  "PyYAML",
  "PySocks",
  "ttp",
  "tld",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/blacknon/pywhois2"
Repository = "https://github.com/blacknon/pywhois2"

[project.scripts]
pywhois2 = "pywhois2:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["pywhois2*"]

[tool.setuptools.package-data]
pywhois2 = ["data/*", "templates/*"]
